+++ lib/providers/home.dart
@@ -2,13 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:miel_work_web/models/chat.dart';
 import 'package:miel_work_web/models/organization.dart';
 import 'package:miel_work_web/models/organization_group.dart';
+import 'package:miel_work_web/models/user.dart';
 import 'package:miel_work_web/services/chat.dart';
+import 'package:miel_work_web/services/log.dart';
 import 'package:miel_work_web/services/organization_group.dart';
 
 class HomeProvider with ChangeNotifier {
   int currentIndex = 0;
+
   final OrganizationGroupService _groupService = OrganizationGroupService();
   final ChatService _chatService = ChatService();
+  final LogService _logService = LogService();
+
   List<OrganizationGroupModel> groups = [];
   OrganizationGroupModel? currentGroup;
 
@@ -46,10 +51,12 @@ class HomeProvider with ChangeNotifier {
   Future<String?> groupCreate({
     required OrganizationModel? organization,
     required String name,
+    required UserModel? loginUser,
   }) async {
     String? error;
     if (organization == null) return 'グループの追加に失敗しました';
     if (name == '') return 'グループ名は必須入力です';
+    if (loginUser == null) return 'グループの追加に失敗しました';
     try {
       String groupId = _groupService.id(organizationId: organization.id);
       _groupService.create({
@@ -86,6 +93,17 @@ class HomeProvider with ChangeNotifier {
         'createdAt': DateTime.now(),
       });
       setGroups(organizationId: organization.id);
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'グループを追加しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = 'グループの追加に失敗しました';
     }
@@ -96,11 +114,13 @@ class HomeProvider with ChangeNotifier {
     required OrganizationModel? organization,
     required OrganizationGroupModel? group,
     required String name,
+    required UserModel? loginUser,
   }) async {
     String? error;
     if (organization == null) return 'グループ名の変更に失敗しました';
     if (group == null) return 'グループ名の変更に失敗しました';
     if (name == '') return 'グループ名は必須入力です';
+    if (loginUser == null) return 'グループ名の変更に失敗しました';
     try {
       _groupService.update({
         'id': group.id,
@@ -118,6 +138,17 @@ class HomeProvider with ChangeNotifier {
         });
       }
       setGroups(organizationId: organization.id);
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'グループ名を変更しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = 'グループ名の変更に失敗しました';
     }
@@ -128,11 +159,13 @@ class HomeProvider with ChangeNotifier {
     required OrganizationModel? organization,
     required OrganizationGroupModel? group,
     required String loginId,
+    required UserModel? loginUser,
   }) async {
     String? error;
     if (organization == null) return 'ログインIDの変更に失敗しました';
     if (group == null) return 'ログインIDの変更に失敗しました';
     if (loginId == '') return 'ログインIDは必須入力です';
+    if (loginUser == null) return 'ログインIDの変更に失敗しました';
     try {
       _groupService.update({
         'id': group.id,
@@ -140,6 +173,17 @@ class HomeProvider with ChangeNotifier {
         'loginId': loginId,
       });
       setGroups(organizationId: organization.id);
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'ログインIDを変更しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = 'ログインIDの変更に失敗しました';
     }
@@ -150,11 +194,13 @@ class HomeProvider with ChangeNotifier {
     required OrganizationModel? organization,
     required OrganizationGroupModel? group,
     required String password,
+    required UserModel? loginUser,
   }) async {
     String? error;
     if (organization == null) return 'パスワードの変更に失敗しました';
     if (group == null) return 'パスワードの変更に失敗しました';
     if (password == '') return 'パスワードは必須入力です';
+    if (loginUser == null) return 'パスワードの変更に失敗しました';
     try {
       _groupService.update({
         'id': group.id,
@@ -162,6 +208,17 @@ class HomeProvider with ChangeNotifier {
         'password': password,
       });
       setGroups(organizationId: organization.id);
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'パスワードを変更しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = 'パスワードの変更に失敗しました';
     }
@@ -171,10 +228,12 @@ class HomeProvider with ChangeNotifier {
   Future<String?> groupDelete({
     required OrganizationModel? organization,
     required OrganizationGroupModel? group,
+    required UserModel? loginUser,
   }) async {
     String? error;
     if (organization == null) return 'グループの削除に失敗しました';
     if (group == null) return 'グループの削除に失敗しました';
+    if (loginUser == null) return 'グループの削除に失敗しました';
     try {
       _groupService.delete({
         'id': group.id,
@@ -190,6 +249,17 @@ class HomeProvider with ChangeNotifier {
         });
       }
       setGroups(organizationId: organization.id);
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'グループを削除しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = 'グループの削除に失敗しました';
     }
