+++ lib/providers/loan.dart
@@ -10,6 +10,7 @@ import 'package:miel_work_web/models/organization.dart';
 import 'package:miel_work_web/models/user.dart';
 import 'package:miel_work_web/services/fm.dart';
 import 'package:miel_work_web/services/loan.dart';
+import 'package:miel_work_web/services/log.dart';
 import 'package:miel_work_web/services/user.dart';
 import 'package:path/path.dart' as p;
 import 'package:signature/signature.dart';
@@ -18,6 +19,7 @@ class LoanProvider with ChangeNotifier {
   final LoanService _loanService = LoanService();
   final UserService _userService = UserService();
   final FmService _fmService = FmService();
+  final LogService _logService = LogService();
 
   Future<String?> create({
     required OrganizationModel? organization,
@@ -90,6 +92,17 @@ class LoanProvider with ChangeNotifier {
           }
         }
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '貸出物を追加しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '貸出物の追加に失敗しました';
     }
@@ -153,6 +166,17 @@ class LoanProvider with ChangeNotifier {
           'itemName': itemName,
         });
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '貸出情報を編集しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '貸出情報の編集に失敗しました';
     }
@@ -190,6 +214,17 @@ class LoanProvider with ChangeNotifier {
         'returnAt': returnAt,
         'returnUser': returnUser,
         'signImage': signImage,
+      });
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '貸出物を返却しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
       });
     } catch (e) {
       error = '貸出物の返却に失敗しました';
@@ -243,6 +278,17 @@ class LoanProvider with ChangeNotifier {
           }
         }
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '貸出物に社内コメントを追記しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '社内コメントの追記に失敗しました';
     }
@@ -251,11 +297,24 @@ class LoanProvider with ChangeNotifier {
 
   Future<String?> delete({
     required LoanModel loan,
+    required UserModel? loginUser,
   }) async {
     String? error;
+    if (loginUser == null) return '貸出情報の削除に失敗しました';
     try {
       _loanService.delete({
         'id': loan.id,
+      });
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': loan.organizationId,
+        'content': '貸出物を削除しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
       });
     } catch (e) {
       error = '貸出情報の削除に失敗しました';
