+++ lib/providers/lost.dart
@@ -9,6 +9,7 @@ import 'package:miel_work_web/models/lost.dart';
 import 'package:miel_work_web/models/organization.dart';
 import 'package:miel_work_web/models/user.dart';
 import 'package:miel_work_web/services/fm.dart';
+import 'package:miel_work_web/services/log.dart';
 import 'package:miel_work_web/services/lost.dart';
 import 'package:miel_work_web/services/user.dart';
 import 'package:path/path.dart' as p;
@@ -18,6 +19,7 @@ class LostProvider with ChangeNotifier {
   final LostService _lostService = LostService();
   final UserService _userService = UserService();
   final FmService _fmService = FmService();
+  final LogService _logService = LogService();
 
   Future<String?> create({
     required OrganizationModel? organization,
@@ -91,6 +93,17 @@ class LostProvider with ChangeNotifier {
           }
         }
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '落とし物を追加しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '落とし物の追加に失敗しました';
     }
@@ -150,6 +163,17 @@ class LostProvider with ChangeNotifier {
           'remarks': remarks,
         });
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '落とし物情報を編集しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '落とし物の編集に失敗しました';
     }
@@ -206,6 +230,17 @@ class LostProvider with ChangeNotifier {
         'returnCustomerIDImage': returnCustomerIDImage,
         'signImage': signImage,
       });
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '落とし物を返却しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '落とし物の返却に失敗しました';
     }
@@ -224,6 +259,17 @@ class LostProvider with ChangeNotifier {
       _lostService.update({
         'id': lost.id,
         'status': 9,
+      });
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '落とし物を破棄しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
       });
     } catch (e) {
       error = '落とし物の破棄に失敗しました';
@@ -277,6 +323,17 @@ class LostProvider with ChangeNotifier {
           }
         }
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '落とし物に社内コメントを追記しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '社内コメントの追記に失敗しました';
     }
@@ -285,11 +342,24 @@ class LostProvider with ChangeNotifier {
 
   Future<String?> delete({
     required LostModel lost,
+    required UserModel? loginUser,
   }) async {
     String? error;
+    if (loginUser == null) return '落とし物情報の削除に失敗しました';
     try {
       _lostService.delete({
         'id': lost.id,
+      });
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': lost.organizationId,
+        'content': '落とし物を削除しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
       });
     } catch (e) {
       error = '落とし物情報の削除に失敗しました';
