+++ lib/providers/notice.dart
@@ -7,6 +7,7 @@ import 'package:miel_work_web/models/organization.dart';
 import 'package:miel_work_web/models/organization_group.dart';
 import 'package:miel_work_web/models/user.dart';
 import 'package:miel_work_web/services/fm.dart';
+import 'package:miel_work_web/services/log.dart';
 import 'package:miel_work_web/services/notice.dart';
 import 'package:miel_work_web/services/user.dart';
 import 'package:path/path.dart' as p;
@@ -15,6 +16,7 @@ class NoticeProvider with ChangeNotifier {
   final NoticeService _noticeService = NoticeService();
   final UserService _userService = UserService();
   final FmService _fmService = FmService();
+  final LogService _logService = LogService();
 
   Future<String?> create({
     required OrganizationModel? organization,
@@ -82,6 +84,17 @@ class NoticeProvider with ChangeNotifier {
           }
         }
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'お知らせを追加しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       print(e);
       error = 'お知らせの追加に失敗しました';
@@ -139,6 +152,17 @@ class NoticeProvider with ChangeNotifier {
           'expirationAt': DateTime.now().add(const Duration(days: 365)),
         });
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'お知らせ情報を編集しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = 'お知らせの編集に失敗しました';
     }
@@ -191,6 +215,17 @@ class NoticeProvider with ChangeNotifier {
           }
         }
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': 'お知らせに社内コメントを追記しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '社内コメントの追記に失敗しました';
     }
@@ -199,8 +234,10 @@ class NoticeProvider with ChangeNotifier {
 
   Future<String?> delete({
     required NoticeModel notice,
+    required UserModel? loginUser,
   }) async {
     String? error;
+    if (loginUser == null) return 'お知らせの削除に失敗しました';
     try {
       _noticeService.delete({
         'id': notice.id,
@@ -212,6 +249,17 @@ class NoticeProvider with ChangeNotifier {
             .child('/${notice.id}${notice.fileExt}')
             .delete();
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': notice.organizationId,
+        'content': 'お知らせを削除しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = 'お知らせの削除に失敗しました';
     }
