+++ lib/providers/plan_garbageman.dart
@@ -4,10 +4,12 @@ import 'package:miel_work_web/models/organization.dart';
 import 'package:miel_work_web/models/plan_garbageman.dart';
 import 'package:miel_work_web/models/plan_garbageman_week.dart';
 import 'package:miel_work_web/models/user.dart';
+import 'package:miel_work_web/services/log.dart';
 import 'package:miel_work_web/services/plan_garbageman.dart';
 
 class PlanGarbagemanProvider with ChangeNotifier {
   final PlanGarbagemanService _garbagemanService = PlanGarbagemanService();
+  final LogService _logService = LogService();
 
   Future<String?> create({
     required OrganizationModel? organization,
@@ -15,10 +17,12 @@ class PlanGarbagemanProvider with ChangeNotifier {
     required DateTime startedAt,
     required DateTime endedAt,
     required String remarks,
+    required UserModel? loginUser,
   }) async {
     String? error;
     if (organization == null) return '勤務予定の追加に失敗しました';
     if (user == null) return '勤務予定の追加に失敗しました';
+    if (loginUser == null) return '勤務予定の追加に失敗しました';
     try {
       String id = _garbagemanService.id();
       _garbagemanService.create({
@@ -32,6 +36,17 @@ class PlanGarbagemanProvider with ChangeNotifier {
         'createdAt': DateTime.now(),
         'expirationAt': startedAt.add(const Duration(days: 365)),
       });
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '勤務予定を追加しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '勤務予定の追加に失敗しました';
     }
@@ -42,11 +57,13 @@ class PlanGarbagemanProvider with ChangeNotifier {
     required OrganizationModel? organization,
     required List<PlanGarbagemanWeekModel> garbagemanWeeks,
     required List<DateTime> days,
+    required UserModel? loginUser,
   }) async {
     String? error;
     if (organization == null) return '1ヵ月分の反映に失敗しました';
     if (garbagemanWeeks.isEmpty) return '1ヵ月分の反映に失敗しました';
     if (days.isEmpty) return '1ヵ月分の反映に失敗しました';
+    if (loginUser == null) return '1ヵ月分の反映に失敗しました';
     try {
       for (final day in days) {
         String week = dateText('E', day);
@@ -81,6 +98,17 @@ class PlanGarbagemanProvider with ChangeNotifier {
           }
         }
       }
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '勤務予定に1ヵ月分を反映しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '1ヵ月分の反映に失敗しました';
     }
@@ -94,10 +122,12 @@ class PlanGarbagemanProvider with ChangeNotifier {
     required DateTime startedAt,
     required DateTime endedAt,
     required String remarks,
+    required UserModel? loginUser,
   }) async {
     String? error;
     if (organization == null) return '勤務予定の編集に失敗しました';
     if (user == null) return '勤務予定の編集に失敗しました';
+    if (loginUser == null) return '勤務予定の編集に失敗しました';
     try {
       _garbagemanService.update({
         'id': garbageman.id,
@@ -109,6 +139,17 @@ class PlanGarbagemanProvider with ChangeNotifier {
         'remarks': remarks,
         'expirationAt': startedAt.add(const Duration(days: 365)),
       });
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': organization.id,
+        'content': '勤務予定を編集しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
+      });
     } catch (e) {
       error = '勤務予定の編集に失敗しました';
     }
@@ -117,11 +158,24 @@ class PlanGarbagemanProvider with ChangeNotifier {
 
   Future<String?> delete({
     required PlanGarbagemanModel garbageman,
+    required UserModel? loginUser,
   }) async {
     String? error;
+    if (loginUser == null) return '勤務予定の削除に失敗しました';
     try {
       _garbagemanService.delete({
         'id': garbageman.id,
+      });
+      //ログ保存
+      String logId = _logService.id();
+      _logService.create({
+        'id': logId,
+        'organizationId': garbageman.organizationId,
+        'content': '勤務予定を削除しました。',
+        'device': 'PC(ブラウザ)',
+        'createdUserId': loginUser.id,
+        'createdUserName': loginUser.name,
+        'createdAt': DateTime.now(),
       });
     } catch (e) {
       error = '勤務予定の削除に失敗しました';
